+++ app/src/test/java/org/simple/clinic/scheduleappointment/ScheduleAppointmentInitTest.kt
@@ -8,6 +8,7 @@ import org.junit.Test
 import org.simple.clinic.newentry.ButtonState
 import org.simple.clinic.overdue.AppointmentConfig
 import org.simple.clinic.overdue.TimeToAppointment
+import org.simple.clinic.summary.AppointmentSheetOpenedFrom.DONE_CLICK
 import org.simple.clinic.util.TestUserClock
 import java.time.LocalDate
 import java.time.Period
@@ -35,7 +36,8 @@ class ScheduleAppointmentInitTest {
         timeToAppointments = appointmentConfig.scheduleAppointmentsIn,
         userClock = clock,
         doneButtonState = ButtonState.SAVED,
-        nextButtonState = NextButtonState.SCHEDULED
+        nextButtonState = NextButtonState.SCHEDULED,
+        openedFrom = DONE_CLICK
     )
 
     initSpec