+++ app/src/test/java/org/simple/clinic/scheduleappointment/ScheduleAppointmentLogicTest.kt
@@ -33,6 +33,7 @@ import org.simple.clinic.overdue.TimeToAppointment.Weeks
 import org.simple.clinic.patient.PatientRepository
 import org.simple.clinic.protocol.Protocol
 import org.simple.clinic.protocol.ProtocolRepository
+import org.simple.clinic.summary.AppointmentSheetOpenedFrom.DONE_CLICK
 import org.simple.clinic.teleconsultlog.teleconsultrecord.TeleconsultRecordRepository
 import org.simple.clinic.util.RxErrorsRule
 import org.simple.clinic.util.TestUserClock
@@ -775,7 +776,8 @@ class ScheduleAppointmentLogicTest {
             timeToAppointments = config.scheduleAppointmentsIn,
             userClock = clock,
             doneButtonState = SAVED,
-            nextButtonState = ButtonState.SCHEDULED
+            nextButtonState = ButtonState.SCHEDULED,
+            openedFrom = DONE_CLICK
         ),
         init = ScheduleAppointmentInit(),
         update = ScheduleAppointmentUpdate(